Implement bottom navigation with fragments and ViewPager2
