+++ CineCraze/app/build.gradle
@@ -51,4 +51,11 @@ dependencies {
     // Room database dependencies
     implementation 'androidx.room:room-runtime:2.6.1'
     annotationProcessor 'androidx.room:room-compiler:2.6.1'
+    
+    // Bubble Navigation for bottom navigation
+    implementation 'com.gauravk.bubblenavigation:bubblenavigation:1.0.7'
+    
+    // ViewPager2 for fragment management
+    implementation 'androidx.viewpager2:viewpager2:1.0.0'
+    implementation 'androidx.fragment:fragment:1.6.2'
 }
