+++ CineCraze/app/scr/main/res/drawable/bg_background_purple.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <solid android:color="#9C27B0" />
+</shape>
