+++ CineCraze/app/scr/main/res/drawable/bg_search.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#33FFFFFF" />
+    <corners android:radius="20dp" />
+    <stroke 
+        android:width="1dp" 
+        android:color="#66FFFFFF" />
+</shape>
