+++ CineCraze/app/scr/main/res/drawable/ic_television_box.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#FFFFFF"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M8.16,3l1.42,1.42 -1.42,1.42L6.58,4.42 8.16,3M12,3l1.42,1.42L12,5.84 10.58,4.42 12,3M15.84,3l1.42,1.42 -1.42,1.42L14.42,4.42 15.84,3M21,6v12c0,1.1 -0.9,2 -2,2H5c-1.1,0 -2,-0.9 -2,-2V6c0,-1.1 0.9,-2 2,-2h14c1.1,0 2,0.9 2,2zM19,8H5v8h14V8z"/>
+</vector>
